requirements.txt
sentence-transformers
pypdf
python-docx
python-pptx
torch